Rename metadata.py
