+++ tests/utils/test_metadata_utils.py
@@ -1,9 +1,9 @@
 import pytest
 
-from hbsir import metadata
+from hbsir import metadata_reader
 from hbsir.utils.metadata_utils import MetadataVersionResolver
 
-inst = metadata.metadatas.instruction
+inst = metadata_reader.metadatas.instruction
 
 simple_metadata = {"key": 1380, "other_key": "other_value"}
